Fix sign-compare warning in PmFromFile::isComment
Use size_t for the loop index when scanning characters before '#'.

Fixes GCC -Wsign-compare:

operators/pm_from_file.cc:36:27: warning: comparison of integer
expressions
of different signedness [-Wsign-compare]

+++ src/operators/pm_from_file.cc
@@ -33,7 +33,7 @@ bool PmFromFile::isComment(const std::string &s) {
     }
     size_t pos = s.find("#");
     if (pos != std::string::npos) {
-        for (int i = 0; i < pos; i++) {
+        for (size_t i = 0; i < pos; i++) {
 	    if (!std::isspace(s[i])) {
 		return false;
 	    }
